feat: timeStamp -> String으로 변경하는 함수 작성 후 적용
